refactor: rename SocialLinks-styling file
