+++ src/components/containers/domains/socials/SocialLinks.js
@@ -1,6 +1,6 @@
 import { View, Text } from 'react-native';
 import Socials from './Socials';
-import styles from './styles/SocialLinksStyles';
+import styles from './SocialLinks.styles';
 import SocialButton from './SocialButton';
 
 export default function SocialLinks() {
