+++ src/components/containers/domains/socials/SocialLinks.styles.js
@@ -1,5 +1,5 @@
 import { StyleSheet } from "react-native";
-import theme from "../../../../../theme";
+import theme from "../../../../theme";
 
 
 const styles = StyleSheet.create({
